requirements.txt
tzdata==2024.1
Django==5.0.1
django-filter==
django-cors-headers==4.3.1
django-filter==24.2
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djangorestframework-xml==2.0.0
drf-yasg==1.21.7
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.7.0
python-dotenv==1.0.1
coverage==7.5.4
stripe==10.2.0